requirements.txt
Flask==2.3.2
twilio==8.8.0
gunicorn==21.2.0
python-dotenv==1.0.1
requests==2.31.0
psycopg2-binary==2.9.7